fix: Use incremental saltNonce for all safe creations
