+++ src/features/counterfactual/utils.ts
@@ -1,25 +1,7 @@
-import { computeNewSafeAddress } from '@/components/new-safe/create/logic'
 import { LATEST_SAFE_VERSION } from '@/config/constants'
-import { isSmartContract } from '@/hooks/wallets/web3'
 import { defaultSafeInfo } from '@/store/safeInfoSlice'
-import type { DeploySafeProps, PredictedSafeProps } from '@safe-global/protocol-kit'
+import type { PredictedSafeProps } from '@safe-global/protocol-kit'
 import { ImplementationVersionState } from '@safe-global/safe-gateway-typescript-sdk'
-import type { BrowserProvider } from 'ethers'
-
-export const getCounterfactualNonce = async (
-  provider: BrowserProvider,
-  props: DeploySafeProps,
-): Promise<string | undefined> => {
-  const safeAddress = await computeNewSafeAddress(provider, props)
-  const isContractDeployed = await isSmartContract(provider, safeAddress)
-
-  // Safe is already deployed so we try the next saltNonce
-  if (isContractDeployed) {
-    return getCounterfactualNonce(provider, { ...props, saltNonce: (Number(props.saltNonce) + 1).toString() })
-  }
-
-  return props.saltNonce
-}
 
 export const getUndeployedSafeInfo = (undeployedSafe: PredictedSafeProps, address: string, chainId: string) => {
   return Promise.resolve({
